Adds evolving game, board, and blocks
Still needs to evolve at the block and lower

a block is a set of three rows

The thought is that each block will contain neighborhoods
and each neighborhood will be a cell and its immediate neighbors.

At that point I should be able to easily evolve a single cell.